Use integers instead of floats for constraint and rule scores Update regression tests now that stonith ops include the node as well   (not just in the attributes) Some tweaks to the processing of colocation constraints for complex   resource types (fixes 861)
CVS patchset: 7459
CVS date: 2005/09/01 11:41:20

Mercurial revision: 8f3f2023ac787e3e355dc910293dca6dbc0d38d3

+++ crm/pengine/pe_utils.h
@@ -1,4 +1,4 @@
-/* $Id: pe_utils.h,v 1.30 2005/08/08 12:09:33 andrew Exp $ */
+/* $Id: pe_utils.h,v 1.31 2005/09/01 11:41:20 andrew Exp $ */
 /* 
  * Copyright (C) 2004 Andrew Beekhof <andrew@beekhof.net>
  * 
@@ -23,7 +23,7 @@
 /* General utilities */
 extern resource_t *pe_find_resource(GListPtr rsc_list, const char *id_rh);
 
-extern float merge_weights(float w1, float w2);
+extern int merge_weights(int w1, int w2);
 
 /* Constraint helper functions */
 extern rsc_colocation_t *invert_constraint(rsc_colocation_t *constraint);
@@ -129,7 +129,7 @@ extern void set_id(crm_data_t *xml_obj, const char *prefix, int child);
 
 extern const char *ordering_type2text(enum pe_ordering type);
 
-extern float char2score(const char *score);
+extern int char2score(const char *score);
 
 
 /* free the various structures */
